Remove example_9 and renamed example_10 to 9
