+++ examples/example_9.py
@@ -6,12 +6,17 @@ from rlpyt.samplers.serial.sampler import SerialSampler
 from rlpyt.samplers.async_.cpu_sampler import AsyncCpuSampler
 from rlpyt.algos.dqn.dqn import DQN
 from rlpyt.agents.dqn.dqn_agent import DqnAgent
+from rlpyt.agents.dqn.atari.atari_dqn_agent import AtariDqnAgent
 from rlpyt.runners.minibatch_rl import MinibatchRlEval, MinibatchRl
 from rlpyt.utils.logging.context import logger_context
 from rlpyt.envs.gym import GymEnvWrapper
 from rlpyt.runners.async_rl import AsyncRlEval
 from rlpyt.utils.tensor import infer_leading_dims, restore_leading_dims
 from rlpyt.utils.wrappers import *
+from rlpyt.envs.gym import make as make_env
+from rlpyt.replays.non_sequence.uniform import UniformReplayBuffer
+from rlpyt.envs.base import EnvSpaces
+from rlpyt.utils.buffer import buffer_to
 
 import torch
 import torch.nn as nn
@@ -19,7 +24,7 @@ import torch.nn.functional as F
 
 import numpy as np
 import gym
-import cv2
+
 
 
 ############################# classes and functions #############################
@@ -33,61 +38,70 @@ class CustomMixin:
 class CustomDqnModel(torch.nn.Module):
     def __init__(
             self,
-            in_channels=1,
-            n_actions=6,
+            observation_shape,
+            output_size,
+            fc_sizes=64
             ):
         super().__init__()
-        self.conv1 = nn.Conv2d(in_channels, 32, kernel_size=4, stride=4)
-        self.conv2 = nn.Conv2d(32, 64, kernel_size=4, stride=2)
-        self.conv3 = nn.Conv2d(64, 64, kernel_size=3, stride=1)
-        self.fc4 = nn.Linear(7 * 7 * 64, 512)
-        # self.fc4 = nn.Linear(10 * 7 * 64, 512)
-        self.fc5 = nn.Linear(512, 256)
-        self.head = nn.Linear(256, n_actions)
+        self._obs_ndim = len(observation_shape)
+        input_shape = observation_shape[0]
+
+        self.base_net = torch.nn.Sequential(
+            torch.nn.Linear(input_shape, fc_sizes),
+            torch.nn.ReLU(),
+            torch.nn.Linear(fc_sizes, fc_sizes),
+            torch.nn.ReLU(),
+            torch.nn.Linear(fc_sizes, output_size),
+        )
+        # self.base_net.apply(self.init_weights)
 
     def forward(self, observation, prev_action, prev_reward):
-        img = observation.type(torch.float)  # Expect torch.uint8 inputs
-        img = img.mul_(1. / 255)  # From [0-255] to [0-1], in place.
-        lead_dim, T, B, img_shape = infer_leading_dims(img, 3)
-        x = img.view(T * B, *img_shape[::-1])
-        # x = img.view(T * B, *img_shape)
-        x = F.relu(self.conv1(x))
-        x = F.relu(self.conv2(x))
-        x = F.relu(self.conv3(x))
-        x = F.relu(self.fc4(x.view(x.size(0), -1)))
-        x = F.relu(self.fc5(x))
-        q = self.head(x)
+        observation = observation.type(torch.float)
+        lead_dim, T, B, obs_shape = infer_leading_dims(observation, self._obs_ndim)
+        obs = observation.view(T * B, -1)
+        q = self.base_net(obs)
         q = restore_leading_dims(q, lead_dim, T, B)
         return q
 
+    def init_weights(self, m):
+        if type(m) == torch.nn.Linear:
+            torch.nn.init.normal_(m.weight)
+            torch.nn.init.zeros_(m.bias)
+
 
 # class CustomDqnAgent(CustomMixin, DqnAgent):
-class CustomDqnAgent(DqnAgent):
+class CustomDqnAgent(CustomMixin, DqnAgent):
     def __init__(self, ModelCls=CustomDqnModel, **kwargs):
         super().__init__(ModelCls=ModelCls, **kwargs)
 
+    @torch.no_grad()
+    def eval_step(self, observation, prev_action, prev_reward):
+        """Computes Q-values for states/observations and selects actions by
+        epsilon-greedy. (no grad)"""
+        # prev_action = self.distribution.to_onehot(prev_action)
+        model_inputs = buffer_to((observation, prev_action, prev_reward),
+                                 device=self.device)
+        q = self.model(*model_inputs)
+        q = q.cpu()
+        action = torch.argmax(q)
+        return action
+
 
 def make_env_custom(*args, **kwargs):
-    env = gym.make('Pong-v0')
-    # env = make_env(env, scale=True)
-    # env = ResizeFrame(env)
-    env = MaxAndSkipEnv(env, skip=4)
-    env = WarpFrame(env)
-    env = FrameStack(env, 4)
+    env = gym.make('CartPole-v0')
     env = GymEnvWrapper(env)
     return env
 
 
 def build_and_train(run_ID=0, cuda_idx=None):
-    # env_id = 'CartPole-v0'
-    env_id = 'Pong-v0'
+    env_id = 'CartPole-v0'
 
-    sampler = AsyncCpuSampler(
-        EnvCls=make_env_custom,
+    sampler = CpuSampler(
+        EnvCls=make_env,
         env_kwargs=dict(id=env_id), #env_config,
         eval_env_kwargs=dict(id=env_id),  #env_config,
-        batch_T=16,  # One time-step per sampler iteration.
-        batch_B=32,  # One environment (i.e. sampler Batch dimension).
+        batch_T=4,  # One time-step per sampler iteration.
+        batch_B=8,  # One environment (i.e. sampler Batch dimension).
         max_decorrelation_steps=100,
         eval_n_envs=2,
         eval_max_steps=int(10e3),
@@ -95,32 +109,27 @@ def build_and_train(run_ID=0, cuda_idx=None):
     )
 
     algo = DQN(
-        replay_ratio = 8,
-        min_steps_learn = 1e4,
-        replay_size = int(1e5)
+        learning_rate=1e-3,
+        replay_ratio=8,
+        batch_size=32,
+        min_steps_learn=32,
+        eps_steps=10e3,
+        replay_size=int(1e3),
+        # double_dqn=True,
+        # target_update_interval=1,
+        # prioritized_replay=True,
+        ReplayBufferCls=UniformReplayBuffer,
     )
 
     agent = CustomDqnAgent()
 
-    affinity = make_affinity(
-        run_slot=0,
-        n_cpu_core=3,  # Use 16 cores across all experiments.
-        n_gpu=1,  # Use 8 gpus across all experiments.
-        sample_gpu_per_run=0,
-        async_sample=True,
-        # hyperthread_offset=24,  # If machine has 24 cores.
-        # n_socket=2,  # Presume CPU socket affinity to lower/upper half GPUs.
-        # gpu_per_run=2,  # How many GPUs to parallelize one run across.
-        # cpu_per_run=1,
-    )
-
-    runner = AsyncRlEval(
+    runner = MinibatchRl(
         algo=algo,
         agent=agent,
         sampler=sampler,
-        n_steps=2e6,
-        log_interval_steps=1e3,
-        affinity=affinity #dict(cuda_idx=cuda_idx, workers_cpus=[0,1,2,4,5,6])
+        n_steps=1e6,
+        log_interval_steps=1e2,
+        affinity=dict(cuda_idx=cuda_idx, workers_cpus=[0, 1, 2, 4, 5, 6])
     )
 
     config = dict(env_id=env_id)
@@ -131,15 +140,56 @@ def build_and_train(run_ID=0, cuda_idx=None):
     with logger_context(log_dir, run_ID, name, config, snapshot_mode='last'):
         runner.train()
 
+
+def evaluate():
+    import time
+    pre_trained_model = '/home/isaac/codes/dd-zero/rlpyt/data/local/2020_04-04_06-52.20/dqn_CartPole-v0/run_0/itr_24713.pkl'
+    data = torch.load(pre_trained_model)
+    agent_state_dict = data['agent_state_dict']
+
+    # for loading pre-trained models see: https://github.com/astooke/rlpyt/issues/69
+    env = gym.make('CartPole-v0')
+
+    agent = CustomDqnAgent(initial_model_state_dict=agent_state_dict['model'])
+
+    env_spaces = EnvSpaces(
+            observation=env.observation_space,
+            action=env.action_space,
+    )
+    agent.initialize(env_spaces)
+    agent.load_state_dict(agent_state_dict['model'])
+
+    obs = env.reset()
+    tot_reward = 0
+    while True:
+        # action = agent.step(torch.tensor(obs, dtype=torch.float32), torch.tensor(0), torch.tensor(0))
+        action = agent.eval_step(torch.tensor(obs, dtype=torch.float32), None, None)
+        a = np.array(action)
+        obs, reward, done, info = env.step(a)
+        tot_reward += reward
+        env.render()
+        time.sleep(0.001)
+        if done:
+            break
+
+    print('reward: ', tot_reward)
+    env.close()
+
+
 if __name__ == "__main__":
     import argparse
 
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('--run_ID', help='run identifier (logging)', type=int, default=0)
     parser.add_argument('--cuda_idx', help='gpu to use ', type=int, default=0)
+    parser.add_argument('--mode', help='train or eval', default='eval')
+
     args = parser.parse_args()
 
-    build_and_train(
-        run_ID=args.run_ID,
-        cuda_idx=args.cuda_idx,
-    )
+    if args.mode == 'train':
+        build_and_train(
+            run_ID=args.run_ID,
+            cuda_idx=args.cuda_idx,
+        )
+    else:
+        evaluate()
